Rename HelmUpgradeBot script
